Drop twelve no-op casts on globals that already have the cast's own type
Six TUs. Each cast restates the declared type of the global it is applied
to (gCheckpointRouteCount s32, gWeirdMenuSaveTimer s8, the three
gSnowBike* s32 tunables, gAndrossProjectileForwardStep and
gAndrossAimedProjectileSpeed, lbl_803DCC44 s32, lbl_803DCF60 s8); a load
already delivers the declared type, so the cast is noise.

Casts on locals are left alone - a local is not truncated into its
register, so a cast there is a real extension.

Every object byte-identical; main.dol OK.

+++ src/dlls/engine/3/3.c
@@ -657,7 +657,7 @@ void Checkpoint_func06(GameObject* obj, CheckpointRouteState* state, int filter)
     }
     else
     {
-        for (i = 0; i < (int)gCheckpointRouteCount; i++)
+        for (i = 0; i < gCheckpointRouteCount; i++)
         {
             e = gCheckpointRouteTable[i].entry;
             if (visited[i] == 0 && (filter == -1 || e->group == filter))
@@ -668,7 +668,7 @@ void Checkpoint_func06(GameObject* obj, CheckpointRouteState* state, int filter)
                 if (nz * nz + (nx * nx + ddy * ddy) < 409600.0f)
                 {
                     stack[count++] = i;
-                    for (j = i; j < (int)gCheckpointRouteCount; j++)
+                    for (j = i; j < gCheckpointRouteCount; j++)
                     {
                         if (filter == gCheckpointRouteTable[j].entry->group)
                         {
